[AI] refactor io_uring into multiple files
